+++ src/papki/vector_file.hpp
@@ -38,7 +38,7 @@ public:
 	 * @brief Current file size.
      * @return current size of the file.
      */
-	size_t size()const override{
+	uint64_t size()const override{
 		return this->data.size();
 	}
